refactor: mv constant
